Seperate the classes
