+++ lib/player_x.rb
@@ -1,0 +1,6 @@
+class PlayerX
+  attr_accessor :symbol
+  def initialize
+    @symbol = 'X'
+  end
+end
