+++ lib/player_y.rb
@@ -1,0 +1,6 @@
+class PlayerY
+  attr_accessor :symbol
+  def initialize
+    @symbol = 'O'
+  end
+end
