+++ lib/winning_condition.rb
@@ -1,25 +1,3 @@
-class Game
-  attr_accessor :choices, :board
-  def initialize
-    @choices = [1, 2, 3, 4, 5, 6, 7, 8, 9]
-    @board = %w[_ _ _ _ _ _ _ _ _]
-  end
-end
-
-class PlayerX
-  attr_accessor :symbol
-  def initialize
-    @symbol = 'X'
-  end
-end
-
-class PlayerY
-  attr_accessor :symbol
-  def initialize
-    @symbol = 'O'
-  end
-end
-
 class WinningCondition
   def condition_checker(board, symbol)
     exitter(symbol) if outside_horizontal(board) || outside_vertical(board) || inside(board) || diagonal(board)
